requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.3
    # via streamlit
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
fastapi==0.124.4
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.13.0
    # via
    #   altair
    #   plotly
numpy==2.3.5
    # via
    #   -r requirements.in
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.0.0
    # via streamlit
plotly==6.5.0
    # via -r requirements.in
protobuf==6.33.2
    # via streamlit
pyarrow==22.0.0
    # via streamlit
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.45
    # via -r requirements.in
starlette==0.50.0
    # via fastapi
streamlit==1.52.1
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.3
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.6.2
    # via requests
uvicorn[standard]==0.38.0
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
